+++ api-tests/relationships/nested-mutations/two-way-backreference/to-many.test.js
@@ -6,10 +6,6 @@ const alphanumGenerator = gen.alphaNumString.notEmpty();
 
 const toStr = items => items.map(item => item.toString());
 
-// `mongodb-memory-server` downloads a binary on first run in CI, which can take
-// a while, so we bump up the timeout here.
-jest.setTimeout(60000);
-
 function setupKeystone(adapterName) {
   return setupServer({
     adapterName,