+++ packages/keystone/lib/Keystone/index.js
@@ -1,3 +1,4 @@
+const { ApolloServer } = require('apollo-server-express');
 const gql = require('graphql-tag');
 const flattenDeep = require('lodash.flattendeep');
 const memoize = require('micro-memoize');
@@ -30,6 +31,7 @@ const {
 const { List } = require('../ListTypes');
 const { DEFAULT_DIST_DIR } = require('../../constants');
 const { CustomProvider, ListAuthProvider, ListCRUDProvider } = require('../providers');
+const { formatError } = require('./format-error');
 
 module.exports = class Keystone {
   constructor({
@@ -239,19 +241,6 @@ module.exports = class Keystone {
     return execute(schema, query, null, context, variables);
   }
 
-  createHTTPContext({ schemaName, req }) {
-    // The GraphQL App uses this method to build up the context required for each incoming query.
-    return {
-      ...this.createContext({
-        schemaName,
-        authentication: { item: req.user, listKey: req.authedListKey },
-        skipAccessControl: false,
-      }),
-      ...this._sessionManager.getContext(req),
-      req,
-    };
-  }
-
   createAuthStrategy(options) {
     const { type: StrategyType, list: listKey, config } = options;
     const { authType } = StrategyType;
@@ -453,6 +442,42 @@ module.exports = class Keystone {
     }
   }
 
+  createApolloServer({ apolloConfig = {}, schemaName, dev }) {
+    // add the Admin GraphQL API
+    const server = new ApolloServer({
+      maxFileSize: 200 * 1024 * 1024,
+      maxFiles: 5,
+      typeDefs: this.getTypeDefs({ schemaName }),
+      resolvers: this.getResolvers({ schemaName }),
+      context: ({ req }) => ({
+        ...this.createContext({
+          schemaName,
+          authentication: { item: req.user, listKey: req.authedListKey },
+          skipAccessControl: false,
+        }),
+        ...this._sessionManager.getContext(req),
+        req,
+      }),
+      ...(process.env.ENGINE_API_KEY
+        ? {
+            engine: { apiKey: process.env.ENGINE_API_KEY },
+            tracing: true,
+          }
+        : {
+            engine: false,
+            // Only enable tracing in dev mode so we can get local debug info, but
+            // don't bother returning that info on prod when the `engine` is
+            // disabled.
+            tracing: dev,
+          }),
+      formatError,
+      ...apolloConfig,
+    });
+    this._schemas[schemaName] = server.schema;
+
+    return server;
+  }
+
   /**
    * @return Promise<null>
    */
@@ -489,13 +514,6 @@ module.exports = class Keystone {
         list => list.views
       ),
     };
-  }
-
-  // It's not Keystone core's responsibility to create an executable schema, but
-  // once one is, Keystone wants to be able to expose the ability to query that
-  // schema, so this function enables other modules to register that function.
-  registerSchema(schemaName, schema) {
-    this._schemas[schemaName] = schema;
   }
 
   getTypeDefs({ schemaName }) {
@@ -623,6 +641,7 @@ module.exports = class Keystone {
     pinoOptions,
     cors = { origin: true, credentials: true },
   } = {}) {
+    this.createApolloServer({ schemaName: 'internal' });
     const middlewares = await this._prepareMiddlewares({ dev, apps, distDir, pinoOptions, cors });
     // These function can't be called after prepare(), so make them throw an error from now on.
     ['extendGraphQLSchema', 'createList', 'createAuthStrategy'].forEach(f => {