+++ packages/mongo-join-builder/tests/relationship.test.js
@@ -3,13 +3,13 @@ const { relationshipTokenizer } = require('../lib/tokenizers');
 describe('Relationship tokenizer', () => {
   test('Uses correct conditions', () => {
     const relationshipConditions = {
-      getRefListAdapter: () => ({ key: 'Bar', model: { collection: { name: 'name' } } }),
+      isRelationship: true,
+      getListByKey: () => ({ adapter: { key: 'Bar', model: { collection: { name: 'name' } } } }),
       field: { many: false },
       path: 'name',
       rel: {},
     };
-    const findFieldAdapterForQuerySegment = jest.fn(() => relationshipConditions);
-    const listAdapter = { findFieldAdapterForQuerySegment, key: 'Foo' };
+    const listAdapter = { key: 'Foo', fieldAdapters: [relationshipConditions] };
 
     expect(relationshipTokenizer(listAdapter, 'name', ['name'], () => 'abc123')).toMatchObject({
       matchTerm: { $expr: { $eq: [{ $size: '$abc123_name' }, 1] } },
@@ -22,15 +22,12 @@ describe('Relationship tokenizer', () => {
         farCollection: 'name',
       },
     });
-    expect(findFieldAdapterForQuerySegment).toHaveBeenCalledTimes(1);
   });
 
   test('returns empty array when no matches found', () => {
-    const findFieldAdapterForQuerySegment = jest.fn(() => {});
-    const listAdapter = { findFieldAdapterForQuerySegment };
+    const listAdapter = { fieldAdapters: [] };
 
     const result = relationshipTokenizer(listAdapter, 'name', ['name'], () => {});
     expect(result).toMatchObject({});
-    expect(findFieldAdapterForQuerySegment).toHaveBeenCalledTimes(1);
   });
 });